+++ __tests__/unit/core-blockchain/stubs/state-storage.ts
@@ -5,6 +5,7 @@ import { Blocks, Interfaces } from "@arkecosystem/crypto";
 export class StateStoreStub implements State.IStateStore {
     public blockchain: any;
     public lastDownloadedBlock: Interfaces.IBlock | undefined;
+    public genesisBlock: Interfaces.IBlock | undefined;
     public blockPing: any;
     public started: boolean;
     public forkedBlock: Interfaces.IBlock | undefined;
@@ -36,6 +37,14 @@ export class StateStoreStub implements State.IStateStore {
         const lastBlock = this.getLastBlock();
 
         return lastBlock ? lastBlock.data.height : 1;
+    }
+
+    public getGenesisBlock(): Interfaces.IBlock {
+        return this.genesisBlock;
+    }
+
+    public setGenesisBlock(block: Interfaces.IBlock): void {
+        this.genesisBlock = block;
     }
 
     public getLastBlock(): Interfaces.IBlock | undefined {