+++ packages/core-api/src/repositories/index.ts
@@ -1,7 +1,0 @@
-import { BlockRepository } from "./blocks";
-import { TransactionsRepository } from "./transactions";
-
-export const blocksRepository = new BlockRepository();
-export const transactionsRepository = new TransactionsRepository();
-
-export { BlockRepository, TransactionsRepository };